Version validation test moved to BlockValidationLogicSpec
